Removes sequences starting with an stop codon
This commit adds the necessary commands to identify sequences starting with an stop
codon and remove these positions from both the join_exons_tmp2.ok.joined.pep and the
join_exons_tmp2.ok.joined files. This commands and the previous ones to remove
sequences with stop codons in the middle are now in the script join_exons_remove_stops.