Fixes blast script to avoid a Compi task failure when there are no results to filter
So far, if the last grep of the blast script does not filter anything because there
are no BLAST results to filter, it results in a Compi task failure because the grep
commmand ends with status 1 and Compi aborts the task. This is not desirable as we
want other files to be analyzed anyway and this commit fixes this behavior.

Also, an "if" is added to the join-exons and predict tasks so that they are skipped
when their respective input files do not exist.