Fixes get_sequences script to always create the ".sequences" file
If the previous script ("final_selection") called in "join_exons" produces an empty
".selection" file, then the "get_sequences" script does not create the ".sequences"
file at all, which causes the "join_exons" script to fail because it expects the file
to exist.

To overcome this issue, the "get_sequences" script creates an empty ".sequences" file
to guarantee that it exists.